+++ packages/protocol/scripts/build.ts
@@ -2,92 +2,15 @@
 
 import Web3V1Celo from '@celo/typechain-target-web3-v1-celo'
 import { execSync } from 'child_process'
-import { existsSync, readJSONSync } from 'fs-extra'
+import fsExtraPkg from 'fs-extra'
 import minimist, { ParsedArgs } from 'minimist'
 import path from 'path'
 import { tsGenerator } from 'ts-generator'
-import { MENTO_PACKAGE, SOLIDITY_08_PACKAGE } from '../contractPackages'
+import { contractPackages, CoreContracts, ImplContracts, Interfaces, ROOT_DIR } from './consts'
 
-const ROOT_DIR = path.normalize(path.join(__dirname, '../'))
+const { existsSync, readJSONSync } = fsExtraPkg
+
 const BUILD_DIR = path.join(ROOT_DIR, process.env.BUILD_DIR ?? './build')
-
-export const ProxyContracts = [
-  'AccountsProxy',
-  'AttestationsProxy',
-  'BlockchainParametersProxy',
-  'DoubleSigningSlasherProxy',
-  'DowntimeSlasherProxy',
-  'ElectionProxy',
-  'EpochRewardsProxy',
-  'EscrowProxy',
-  'FederatedAttestationsProxy',
-  'FeeHandlerProxy',
-  'MentoFeeHandlerSellerProxy',
-  'FeeCurrencyWhitelistProxy',
-  'GoldTokenProxy',
-  'GovernanceApproverMultiSigProxy',
-  'GovernanceProxy',
-  'LockedGoldProxy',
-  'MetaTransactionWalletProxy',
-  'MetaTransactionWalletDeployerProxy',
-  'OdisPaymentsProxy',
-  'RegistryProxy',
-  'SortedOraclesProxy',
-  'UniswapFeeHandlerSellerProxy',
-]
-
-export const CoreContracts = [
-  // common
-  'Accounts',
-  'GasPriceMinimum',
-  'FeeHandler',
-  'MentoFeeHandlerSeller',
-  'UniswapFeeHandlerSeller',
-  'FeeCurrencyWhitelist',
-  'GoldToken',
-  'MetaTransactionWallet',
-  'MetaTransactionWalletDeployer',
-  'MultiSig',
-  'Registry',
-  'Freezer',
-  'MetaTransactionWallet',
-
-  // governance
-  'Election',
-  'EpochRewards',
-  'Governance',
-  'GovernanceApproverMultiSig',
-  'BlockchainParameters',
-  'DoubleSigningSlasher',
-  'DowntimeSlasher',
-  'LockedGold',
-  'Validators',
-  'ReleaseGold',
-
-  // identity
-  'Attestations',
-  'Escrow',
-  'FederatedAttestations',
-  'Random',
-  'OdisPayments',
-
-  // stability
-  'SortedOracles',
-]
-
-const OtherContracts = [
-  'Proxy',
-  'Migrations',
-  // abstract
-  'Initializable',
-  'UsingRegistry',
-]
-
-const contractPackages = [MENTO_PACKAGE, SOLIDITY_08_PACKAGE]
-
-const Interfaces = ['ICeloToken', 'IERC20', 'ICeloVersionedContract']
-
-export const ImplContracts = OtherContracts.concat(ProxyContracts).concat(CoreContracts)
 
 function exec(cmd: string) {
   return execSync(cmd, { cwd: ROOT_DIR, stdio: 'inherit' })
@@ -97,7 +20,7 @@ function hasEmptyBytecode(contract: any) {
   return contract.bytecode === '0x'
 }
 
-function compile(outdir: string) {
+function compile({ coreContractsOnly, solidity: outdir }: BuildTargets) {
   console.log(`protocol: Compiling solidity to ${outdir}`)
 
   // the reason to generate a different folder is to avoid path collisions, which could be very dangerous
@@ -125,8 +48,9 @@ function compile(outdir: string) {
     `yarn run --silent truffle compile --contracts_directory="./contracts/" --build_directory=${outdir}`
   )
 
+  const contracts = coreContractsOnly ? CoreContracts : ImplContracts
   // check that there were no errors
-  for (const contractName of ImplContracts) {
+  for (const contractName of contracts) {
     try {
       // This is issuing a warning: https://github.com/celo-org/celo-monorepo/issues/10564
       const fileStr = readJSONSync(`${outdir}/contracts/${contractName}.json`)
@@ -145,7 +69,7 @@ function compile(outdir: string) {
   }
 }
 
-function generateFilesForTruffle(outdir: string) {
+function generateFilesForTruffle({ coreContractsOnly, truffleTypes: outdir }: BuildTargets) {
   // tslint:disable-next-line
   for (let externalContractPackage of contractPackages) {
     const outdirExternal = outdir + '-' + externalContractPackage.name
@@ -161,16 +85,34 @@ function generateFilesForTruffle(outdir: string) {
 
   console.log(`protocol: Generating Truffle Types to ${outdir}`)
   exec(`rm -rf "${outdir}"`)
-  const globPattern = `${BUILD_DIR}/contracts/*.json`
+
+  const globPattern = coreContractsOnly
+    ? `${BUILD_DIR}/contracts/@(${CoreContracts.join('|')}).json`
+    : `${BUILD_DIR}/contracts/*.json`
+
   exec(`yarn run --silent typechain --target=truffle --outDir "${outdir}" "${globPattern}"`)
 }
 
-async function generateFilesForContractKit(outdir: string) {
+function generateFilesForEthers({ coreContractsOnly, ethersTypes: outdir }: BuildTargets) {
+  console.log(`protocol: Generating Ethers Types to ${outdir}`)
+  exec(`rm -rf "${outdir}"`)
+
+  const contractKitContracts = coreContractsOnly
+    ? CoreContracts
+    : CoreContracts.concat('Proxy').concat(Interfaces)
+  const globPattern = `${BUILD_DIR}/contracts/@(${contractKitContracts.join('|')}).json`
+
+  exec(`yarn run --silent typechain --target=ethers-v5 --outDir "${outdir}" "${globPattern}"`)
+}
+
+async function generateFilesForContractKit({ coreContractsOnly, web3Types: outdir }: BuildTargets) {
   console.log(`protocol: Generating Web3 Types to ${outdir}`)
   exec(`rm -rf ${outdir}`)
   const relativePath = path.relative(ROOT_DIR, outdir)
 
-  const contractKitContracts = CoreContracts.concat('Proxy').concat(Interfaces)
+  const contractKitContracts = coreContractsOnly
+    ? CoreContracts
+    : CoreContracts.concat('Proxy').concat(Interfaces)
 
   const globPattern = `${BUILD_DIR}/contracts/@(${contractKitContracts.join('|')}).json`
 
@@ -206,27 +148,35 @@ async function generateFilesForContractKit(outdir: string) {
 }
 
 const _buildTargets: ParsedArgs = {
-  _: [],
+  _: [] as string[],
   solidity: undefined,
   truffleTypes: undefined,
   web3Types: undefined,
+  ethersTypes: undefined,
+} as const
+type BuildTargets = Record<keyof typeof _buildTargets, string> & {
+  coreContractsOnly: boolean
 }
 
-async function main(buildTargets: ParsedArgs) {
+async function main(buildTargets: BuildTargets) {
   if (buildTargets.solidity) {
-    compile(buildTargets.solidity)
+    compile(buildTargets)
   }
   if (buildTargets.truffleTypes) {
-    generateFilesForTruffle(buildTargets.truffleTypes)
+    generateFilesForTruffle(buildTargets)
+  }
+  if (buildTargets.ethersTypes) {
+    generateFilesForEthers(buildTargets)
   }
   if (buildTargets.web3Types) {
-    await generateFilesForContractKit(buildTargets.web3Types)
+    await generateFilesForContractKit(buildTargets)
   }
 }
 
 const argv = minimist(process.argv.slice(2), {
   string: Object.keys(_buildTargets),
-})
+  boolean: ['coreContractsOnly'],
+}) as unknown as BuildTargets
 
 main(argv).catch((err) => {
   console.error(err)